+++ sdk/highlight-run/src/client/utils/sessionStorage/highlightSession.ts
@@ -112,16 +112,10 @@ function pruneSessionData(keepKey: string): void {
 				const sessionData = JSON.parse(
 					getItem(key) || '{}',
 				) as SessionData
-				if (sessionData.lastPushTime === undefined) {
-					internalLogOnce(
-						'highlightSession',
-						'pruneSessionData',
-						'error',
-						`data for key ${key} is not session data`,
-					)
-				} else if (
+				if (
+					sessionData.lastPushTime !== undefined &&
 					Date.now() - sessionData.lastPushTime >=
-					SESSION_PUSH_THRESHOLD
+						SESSION_PUSH_THRESHOLD
 				) {
 					internalLogOnce(
 						'highlightSession',